MultiDocumentApp basics (base class for Application, Applications override methods to provide shells for new documents or to load existing documents, Application menu keeps track of open shells).
Todo: Wizard Shell for Operating Systems that don't support Window-less
Application menus.

ProgressBar Dialog.

Todo: Improve Layout/Functionality.

AutoShell: Shell that implements AutoCloseable and can be used in a 
try-block to ensure automatic close after it is not needed anymore.
(Used for invisible shells that are needed as parent for dialogs
when the application has no window yet).