Moved the position/frame handling code into VideoEditor (which serves as a mediator between different UI components that either want to display frame-based information or that want to modify the current position, like the Timeline and the VideoView).
Created an Animator class that could take over the animation code from
the VideoView (for video playback), but somehow the same code as in the
View itself turned out to be laggy when moved to the external class.
Needs to be investigated. VideoView for now continues to use the local
animation/playback code.