Event capturing fixes:
Capturing events (LayerContainer.capture(Layer)) now resets the captive
of the previous capturing Layer to null (or else it can happen in some
cases that the layer that was previously the capturing target won't
receive future events as the event handling routine of LayerCanvas etc
checks whether the captive is equal to the current captive or null).

Resetting the captive in the capture() method that way also made the
resetCaptive() function, that was used by handles when the dragging was
initiated programmatically, unnecessary.

Added additional 'getParentContainer()' method to the LayerContainer
interface, to allow following up the parent-chain from a Layer to the
root (as getParent() returned a LayerContainer and not a Layer). Had to
be renamed to 'getParentContainer()' due to naming conflict with
'getParent()' from the SWT Canvas (which returns a Composite).